+++ Views/CookWithPage.xaml.cs
@@ -1,0 +1,20 @@
+using Popote.ViewModels;
+
+namespace Popote.Views;
+
+public partial class CookWithPage : ContentPage
+{
+    private readonly CookWithViewModel _vm;
+
+    public CookWithPage(CookWithViewModel vm)
+    {
+        InitializeComponent();
+        BindingContext = _vm = vm;
+    }
+
+    protected override async void OnAppearing()
+    {
+        base.OnAppearing();
+        await _vm.LoadCommand.ExecuteAsync(null);
+    }
+}